LVRUBYM-356: Add spec helper method for student attributes
